move AppConfig.ts
