+++ src/webapp/components/app/AppConfig.ts
@@ -1,5 +1,5 @@
 import { FeedbackOptions } from "@eyeseetea/feedback-component";
-import {JsonConfig} from "./data/ConfigWebRepository";
+import {JsonConfig} from "../../../data/ConfigWebRepository";
 
 export interface AppConfig extends JsonConfig {
     appearance: {
